Move sniffer to its own module
