+++ ooni/tests/test_director.py
@@ -70,14 +70,14 @@ class TestStartSniffing(unittest.TestCase):
 
     def test_start_sniffing_once(self):
         with patch('ooni.settings.config.scapyFactory') as mock_scapy_factory:
-            with patch('ooni.utils.txscapy.ScapySniffer') as mock_scapy_sniffer:
+            with patch('ooni.sniffer.ScapySniffer') as mock_scapy_sniffer:
                 self.director.startSniffing(self.testDetails)
                 sniffer = mock_scapy_sniffer.return_value
                 mock_scapy_factory.registerProtocol.assert_called_once_with(sniffer)
 
     def test_start_sniffing_twice(self):
         with patch('ooni.settings.config.scapyFactory') as mock_scapy_factory:
-            with patch('ooni.utils.txscapy.ScapySniffer') as mock_scapy_sniffer:
+            with patch('ooni.sniffer.ScapySniffer') as mock_scapy_sniffer:
                 sniffer = mock_scapy_sniffer.return_value
                 sniffer.pcapwriter.filename = 'foo1_filename'
                 self.director.startSniffing(self.testDetails)
@@ -87,7 +87,7 @@ class TestStartSniffing(unittest.TestCase):
                 'test_name': 'bar',
                 'start_time': time.time()
             }
-            with patch('ooni.utils.txscapy.ScapySniffer') as mock_scapy_sniffer:
+            with patch('ooni.sniffer.ScapySniffer') as mock_scapy_sniffer:
                 sniffer = mock_scapy_sniffer.return_value
                 sniffer.pcapwriter.filename = 'foo2_filename'
                 self.director.startSniffing(self.testDetails)
@@ -95,7 +95,7 @@ class TestStartSniffing(unittest.TestCase):
 
     def test_measurement_succeeded(self):
         with patch('ooni.settings.config.scapyFactory') as mock_scapy_factory:
-            with patch('ooni.utils.txscapy.ScapySniffer') as mock_scapy_sniffer:
+            with patch('ooni.sniffer.ScapySniffer') as mock_scapy_sniffer:
                 self.director.startSniffing(self.testDetails)
                 self.assertEqual(len(self.director.sniffers), 1)
                 measurement = MagicMock()
