+++ ooni/tests/test_txscapy.py
@@ -1,5 +1,4 @@
 import os
-import re
 
 from mock import MagicMock
 from twisted.internet import defer
@@ -24,12 +23,7 @@ class TestTxScapy(unittest.TestCase):
         for filename in self.filenames:
             os.unlink(filename)
 
-    def test_pcapdnet_installed(self):
-        assert txscapy.pcapdnet_installed() is True
-
     def test_send_packet_no_answer(self):
-        from scapy.all import IP, TCP
-
         sender = txscapy.ScapySender()
         self.scapy_factory.registerProtocol(sender)
         packet = IP(dst='8.8.8.8') / TCP(dport=53)
@@ -39,8 +33,6 @@ class TestTxScapy(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_send_packet_with_answer(self):
-        from scapy.all import IP, TCP
-
         sender = txscapy.ScapySender()
         self.scapy_factory.registerProtocol(sender)
 
@@ -61,80 +53,3 @@ class TestTxScapy(unittest.TestCase):
     def test_get_addresses(self):
         addresses = txscapy.getAddresses()
         assert isinstance(addresses, list)
-
-    def test_sniffer_http_url(self):
-        pcap_filename = 'sniffer.pcap'
-        self.filenames.append(pcap_filename)
-        packet = IP(src='10.0.2.1', dst='10.0.2.2') / TCP(dport=80, sport=8080)
-        packet.payload.payload.original = "GET /something/fancy HTTP/1.0\r\n\r\n"
-        sniffer = txscapy.ScapySniffer(pcap_filename)
-        sniffer.filters.append({'http_url': 'www.torproject.org/something/fancy'})
-        sniffer.packetReceived(packet)
-
-        self.assertEqual(len(sniffer._conns), 1)
-        self.assertEqual(sniffer._conns[0]['src'], '10.0.2.1')
-        self.assertEqual(sniffer._conns[0]['dst'], '10.0.2.2')
-        self.assertEqual(sniffer._conns[0]['dport'], 80)
-        self.assertEqual(sniffer._conns[0]['sport'], 8080)
-        size = os.stat(pcap_filename).st_size
-        self.assertGreater(size, 0)
-
-        packet = IP(src='10.0.2.2', dst='10.0.2.1') / TCP(dport=8080, sport=81)
-        packet.payload.payload.original = "GET /something/weird HTTP/1.0\r\n\r\n"
-        sniffer.packetReceived(packet)
-        self.assertEqual(size, os.stat(pcap_filename).st_size)
-
-        packet = IP(src='10.0.2.2', dst='10.0.2.1') / TCP(dport=8080, sport=80)
-        sniffer.packetReceived(packet)
-        self.assertGreater(os.stat(pcap_filename).st_size, size)
-
-    def test_sniffer_http_url_with_http(self):
-        pcap_filename = 'sniffer.pcap'
-        self.filenames.append(pcap_filename)
-        packet = IP(src='10.0.2.2', dst='10.0.2.1') / TCP(dport=8080, sport=80)
-        packet.payload.payload.original = "GET /something/fancy HTTP/1.0\r\n\r\n"
-        sniffer = txscapy.ScapySniffer(pcap_filename)
-        sniffer.filters.append({'http_url': 'http://www.torproject.org/something/fancy'})
-        sniffer.packetReceived(packet)
-
-        self.assertEqual(len(sniffer._conns), 1)
-        size = os.stat(pcap_filename).st_size
-        self.assertGreater(size, 0)
-
-    def test_sniffer_http_url_root(self):
-        pcap_filename = 'sniffer.pcap'
-        self.filenames.append(pcap_filename)
-        packet = IP(src='10.0.2.2', dst='10.0.2.1') / TCP(dport=8080, sport=80)
-        packet.payload.payload.original = "GET / HTTP/1.0\r\n\r\n"
-        sniffer = txscapy.ScapySniffer(pcap_filename)
-        sniffer.filters.append({'http_url': 'www.torproject.org'})
-        sniffer.packetReceived(packet)
-
-        size = os.stat(pcap_filename).st_size
-        self.assertGreater(size, 0)
-
-        packet = IP() / TCP()
-        packet = IP(src='10.0.2.2', dst='10.0.2.1') / TCP(dport=8080, sport=80)
-        packet.payload.payload.original = "DUMB / HTTP/1.0\r\n\r\n"
-        sniffer.packetReceived(packet)
-        self.assertEqual(os.stat(pcap_filename).st_size, size)
-
-    def test_sniffer_http_url_ip_dst(self):
-        pcap_filename = 'sniffer.pcap'
-        self.filenames.append(pcap_filename)
-        packet = IP(src='10.0.2.1', dst='10.0.2.2') / TCP(dport=8080, sport=80)
-        packet.payload.payload.original = "GET / HTTP/1.0\r\n\r\n"
-        sniffer = txscapy.ScapySniffer(pcap_filename)
-        sniffer.filters.append({'http_url': 'http://10.0.2.2'})
-        sniffer.packetReceived(packet)
-
-        size = os.stat(pcap_filename).st_size
-        self.assertGreater(size, 0)
-
-    def test_sniffer_regex_ip(self):
-        pcap_filename = 'sniffer.pcap'
-        self.filenames.append(pcap_filename)
-        sniffer = txscapy.ScapySniffer(pcap_filename)
-        self.assertIsNotNone(re.match(sniffer.ip_regex, '110.0.2.2'))
-        self.assertIsNone(re.match(sniffer.ip_regex, '1.1.1.1.1'))
-        self.assertIsNone(re.match(sniffer.ip_regex, '1111.1.1.1'))
